Add type ascriptions

+++ t/sql/02/05-with-params.sql
@@ -1,2 +1,2 @@
 -- sub base-query(Int $a, Int $b --> $)
-SELECT $a::int + $b::int;
+SELECT $a + $b;
